Add emdash to allowed characters.
I had an audiobook where this came out weird in the filenames.  Adding emdash seemed the easiest thing to do.  I also removed the space before the dash, because spaces in filenames make life more difficult than it needs to be.  I could also rewrite this to replace the emdash with a dash.